requirements.txt
flask==3.1.2
flask-socketio==5.5.1
opencv-contrib-python==4.12.0.88
fpdf2==2.8.4
numpy==2.0.2
pandas==2.2.3
